requirements.txt
Django==5.2.5
Pillow==11.3.0
reportlab==4.4.3
google-generativeai==0.8.3
requests==2.31.0
numpy==2.3.2
matplotlib==3.10.5